requirements.txt
torchvision
numpy
torch
Pillow
